+++ app/assets/locales/locale-en.json
@@ -1334,6 +1334,7 @@
         "advanced_report": "Advanced report",
         "console_report": "Console report",
         "copy_to_clipboard": "Copy to clipboard",
+        "hide": "Hide",
         "title": "Error report",
         "report_description": "This report collects the last entries in the console log and takes a screenshot of the current view before the modal is opened.",
         "link_to_github": "Please report the issue you have found to GitHub:",
